examples/google-basic/pyproject.toml
[project]
name = "example-google-basic"
version = "0.0.0"
description = "Basic Google Gemini agent example"
requires-python = ">=3.10"
dependencies = [
    "reminix-runtime",
    "reminix-google",
    "google-genai>=0.6.0",
    "python-dotenv>=1.0.0",
]

[tool.uv.sources]
reminix-runtime = { workspace = true }
reminix-google = { workspace = true }
